Home page and spinning wheel game.
In the login screen of my second-year final project, which I completed with my teammate, the user has 3 attempts to correctly enter their username and password. On the main page where the game is located, clicking the button in the center of the spinning wheel causes the wheel to spin, and the resulting value affects the bank account. The bank values are stored on localhost.